+++ fpy2/interpret/native.py
@@ -1,141 +1,128 @@
 """
-FPy runtime backed by the Titanic library.
+FPy runtime backed by the Python runtime.
 """
+
+import math
+import titanfp.titanic.gmpmath as gmpmath
 
 from typing import Any, Callable, Optional, Sequence, TypeAlias
 
 from titanfp.arithmetic.evalctx import EvalCtx, determine_ctx
-from titanfp.arithmetic.ieee754 import ieee_ctx
-from titanfp.arithmetic.mpmf import MPMF
+from titanfp.arithmetic.ieee754 import Float, IEEECtx, ieee_ctx
 from titanfp.titanic.digital import Digital
-from titanfp.titanic.ndarray import NDArray
-from titanfp.titanic.ops import OP
-import titanfp.titanic.gmpmath as gmpmath
+from titanfp.titanic.ops import RM
 
-from ..common import ExprTraceEntry
-from ..env import ForeignEnv
-from ..function import Interpreter, Function, FunctionReturnException
-from ...ir import *
+from ..runtime.function import Interpreter, Function, FunctionReturnException
+from ..runtime.env import ForeignEnv
+from ..ir import *
 
-ScalarVal: TypeAlias = bool | Digital
-"""Type of scalar values in FPy programs."""
-TensorVal: TypeAlias = NDArray
-"""Type of tensor values in FPy programs."""
+def _safe_div(x: float, y: float):
+    if y == 0:
+        if x == 0:
+            return math.nan
+        else:
+            return math.copysign(math.inf, x)
+    else:
+        return x / y
 
-ScalarArg: TypeAlias = ScalarVal | str | int | float
-"""Type of scalar arguments in FPy programs; includes native Python types"""
-TensorArg: TypeAlias = NDArray | tuple | list
-"""Type of tensor arguments in FPy programs; includes native Python types"""
-
-def _isinf(x: MPMF) -> bool:
-    return x.isinf
-
-def _isnan(x: MPMF) -> bool:
-    return x.isnan
-
+ScalarVal: TypeAlias = bool | float
+"""Type of scalar values."""
+TensorVal: TypeAlias = tuple
+"""Type of tensor values."""
 _method_table: dict[str, Callable[..., Any]] = {
-    '+': MPMF.add,
-    '-': MPMF.sub,
-    '*': MPMF.mul,
-    '/': MPMF.div,
-    'fabs': MPMF.fabs,
-    'sqrt': MPMF.sqrt,
-    'fma': MPMF.fma,
-    'neg': MPMF.neg,
-    'copysign': MPMF.copysign,
-    'fdim': MPMF.fdim,
-    'fmax': MPMF.fmax,
-    'fmin': MPMF.fmin,
-    'fmod': MPMF.fmod,
-    'remainder': MPMF.remainder,
-    'hypot': MPMF.hypot,
-    'cbrt': MPMF.cbrt,
-    'ceil': MPMF.ceil,
-    'floor': MPMF.floor,
-    'nearbyint': MPMF.nearbyint,
-    'round': MPMF.round,
-    'trunc': MPMF.trunc,
-    'acos': MPMF.acos,
-    'asin': MPMF.asin,
-    'atan': MPMF.atan,
-    'atan2': MPMF.atan2,
-    'cos': MPMF.cos,
-    'sin': MPMF.sin,
-    'tan': MPMF.tan,
-    'acosh': MPMF.acosh,
-    'asinh': MPMF.asinh,
-    'atanh': MPMF.atanh,
-    'cosh': MPMF.cosh,
-    'sinh': MPMF.sinh,
-    'tanh': MPMF.tanh,
-    'exp': MPMF.exp_,
-    'exp2': MPMF.exp2,
-    'expm1': MPMF.expm1,
-    'log': MPMF.log,
-    'log10': MPMF.log10,
-    'log1p': MPMF.log1p,
-    'log2': MPMF.log2,
-    'pow': MPMF.pow,
-    'erf': MPMF.erf,
-    'erfc': MPMF.erfc,
-    'lgamma': MPMF.lgamma,
-    'tgamma': MPMF.tgamma,
-    'isfinite': MPMF.isfinite,
-    'isinf': _isinf,
-    'isnan': _isnan,
-    'isnormal': MPMF.isnormal,
-    'signbit': MPMF.signbit,
+    '+': lambda x, y: x + y,
+    '-': lambda x, y: x - y,
+    '*': lambda x, y: x * y,
+    '/': _safe_div,
+    'fabs': math.fabs,
+    'sqrt': math.sqrt,
+    # TODO: only available in Python 3.13
+    # 'fma': math.fma,
+    'neg': lambda x: -x,
+    'copysign': math.copysign,
+    'fdim': lambda x, y: max(x - y, 0),
+    'fmax': max,
+    'fmin': min,
+    'fmod': math.fmod,
+    'remainder': math.remainder,
+    'hypot': math.hypot,
+    'cbrt': math.cbrt,
+    'ceil': math.ceil,
+    'floor': math.floor,
+    'nearbyint': lambda x: round(x),
+    'round': round,
+    'trunc': math.trunc,
+    'acos': math.acos,
+    'asin': math.asin,
+    'atan': math.atan,
+    'atan2': math.atan2,
+    'cos': math.cos,
+    'sin': math.sin,
+    'tan': math.tan,
+    'acosh': math.acosh,
+    'asinh': math.asinh,
+    'atanh': math.atanh,
+    'cosh': math.cosh,
+    'sinh': math.sinh,
+    'tanh': math.tanh,
+    'exp': math.exp,
+    'exp2': lambda x: 2 ** x,
+    'expm1': math.expm1,
+    'log': math.log,
+    'log10': math.log10,
+    'log1p': math.log1p,
+    'log2': math.log2,
+    'pow': math.pow,
+    'erf': math.erf,
+    'erfc': math.erfc,
+    'lgamma': math.lgamma,
+    'tgamma': math.gamma,
+    'isfinite': math.isfinite,
+    'isinf': math.isinf,
+    'isnan': math.isnan,
+    'isnormal': lambda x: math.isfinite(x) and x != 0,
+    'signbit': lambda x: math.copysign(1, x) < 0,
 }
 
 _Env: TypeAlias = dict[NamedId, ScalarVal | TensorVal]
 
 class _Interpreter(ReduceVisitor):
-    """Single-use interpreter for a function"""
+    """Single-use interpreter for a function."""
 
     foreign: ForeignEnv
     """foreign environment"""
-    override_ctx: Optional[EvalCtx]
-    """optional overriding context"""
     env: _Env
-    """Environment mapping variable names to values"""
-    trace: list[ExprTraceEntry]
-    """expression trace"""
-    enable_trace: bool
-    """expression tracing enabled?"""
+    """environment mapping variable names to values"""
 
-    def __init__(
-        self, 
-        foreign: ForeignEnv,
-        *,
-        override_ctx: Optional[EvalCtx] = None,
-        env: Optional[_Env] = None,
-        enable_trace: bool = False
-    ):
+    def __init__(self, foreign: ForeignEnv, *, env: Optional[_Env] = None):
         if env is None:
             env = {}
 
         self.foreign = foreign
-        self.override_ctx = override_ctx
         self.env = env
-        self.trace = []
-        self.enable_trace = enable_trace
 
-    def _eval_ctx(self, ctx: EvalCtx):
-        if self.override_ctx is None:
-            return ctx
-        else:
-            return self.override_ctx
+    def _is_python_ctx(self, ctx: EvalCtx):
+        return (
+            isinstance(ctx, IEEECtx)
+            and ctx.es == 11
+            and ctx.nbits == 64
+            or ctx.rm == RM.RNE
+        )
 
-    # TODO: what are the semantics of arguments
-    def _arg_to_mpmf(self, arg: Any, ctx: EvalCtx):
-        if isinstance(arg, int | float):
-            return MPMF(x=arg, ctx=ctx)
+    def _arg_to_float(self, arg: Any):
+        if isinstance(arg, str | int | float):
+            return float(arg)
         elif isinstance(arg, Digital):
-            return MPMF(x=arg, ctx=ctx)
+            return float(arg)
         elif isinstance(arg, tuple | list):
-            return NDArray([self._arg_to_mpmf(x, ctx) for x in arg])
+            raise NotImplementedError()
         else:
             return arg
+
+    def _lookup(self, name: NamedId):
+        if name not in self.env:
+            raise RuntimeError(f'unbound variable {name}')
+        return self.env[name]
 
     def eval(
         self,
@@ -143,26 +130,29 @@ class _Interpreter(ReduceVisitor):
         args: Sequence[Any],
         ctx: Optional[EvalCtx] = None
     ):
-        # check arity
         args = tuple(args)
         if len(args) != len(func.args):
             raise TypeError(f'Expected {len(func.args)} arguments, got {len(args)}')
 
-        # determine context if `None` is specified
+        # determine context
         if ctx is None:
             ctx = ieee_ctx(11, 64)
         ctx = determine_ctx(ctx, func.ctx)
 
-        # process arguments and add to environment
+        # Python only has doubles
+        if not self._is_python_ctx(ctx):
+            raise ValueError(f'Unsupported context {ctx}; Python only has doubles')
+
+        # bind arguments
         for val, arg in zip(args, func.args):
             match arg.ty:
                 case AnyType():
-                    x = self._arg_to_mpmf(val, ctx)
+                    x = self._arg_to_float(val)
                     if isinstance(arg.name, NamedId):
                         self.env[arg.name] = x
                 case RealType():
-                    x = self._arg_to_mpmf(val, ctx)
-                    if not isinstance(x, Digital):
+                    x = self._arg_to_float(val)
+                    if not isinstance(x, float):
                         raise NotImplementedError(f'argument is a scalar, got data {val}')
                     if isinstance(arg.name, NamedId):
                         self.env[arg.name] = x
@@ -171,91 +161,73 @@ class _Interpreter(ReduceVisitor):
 
         # process free variables
         for var in func.free_vars:
-            x = self._arg_to_mpmf(self.foreign[var.base], ctx)
+            x = self._arg_to_float(self.foreign[var.base])
             self.env[var] = x
 
-        # evaluation
+        # evaluate the body
         try:
             self._visit_block(func.body, ctx)
             raise RuntimeError('no return statement encountered')
         except FunctionReturnException as e:
             return e.value
 
-    def _lookup(self, name: NamedId):
-        if name not in self.env:
-            raise RuntimeError(f'unbound variable {name}')
-        return self.env[name]
-
     def _visit_var(self, e: Var, ctx: EvalCtx):
         return self._lookup(e.name)
 
-    def _visit_bool(self, e: Bool, ctx: Any):
+    def _visit_bool(self, e: Bool, ctx: EvalCtx):
         return e.val
 
     def _visit_decnum(self, e: Decnum, ctx: EvalCtx):
-        ctx = self._eval_ctx(ctx)
-        return MPMF(x=e.val, ctx=ctx)
-
-    def _visit_integer(self, e: Integer, ctx: EvalCtx):
-        ctx = self._eval_ctx(ctx)
-        x = Digital(m=e.val, exp=0, inexact=False)
-        return MPMF._round_to_context(x, ctx=ctx)
+        return float(e.val)
 
     def _visit_hexnum(self, e: Hexnum, ctx: EvalCtx):
-        return MPMF(x=e.val, ctx=ctx)
+        return float.fromhex(e.val)
+
+    def _visit_integer(self, e: Integer, ctx: EvalCtx):
+        return float(e.val)
 
     def _visit_rational(self, e: Rational, ctx: EvalCtx):
-        ctx = self._eval_ctx(ctx)
-        p = Digital(m=e.p, exp=0, inexact=False)
-        q = Digital(m=e.q, exp=0, inexact=False)
-        x = gmpmath.compute(OP.div, p, q, prec=ctx.p)
-        return MPMF._round_to_context(x, ctx=ctx)
+        return e.p / e.q
 
     def _visit_constant(self, e: Constant, ctx: EvalCtx):
-        ctx = self._eval_ctx(ctx)
+        # rely on Titanic for this
         x = gmpmath.compute_constant(e.val, prec=ctx.p)
-        return MPMF._round_to_context(x, ctx=ctx)
+        d = Float._round_to_context(x, ctx=ctx)
+        return float(d)
 
     def _visit_digits(self, e: Digits, ctx: EvalCtx):
-        ctx = self._eval_ctx(ctx)
+        # rely on Titanic for this
         x = gmpmath.compute_digits(e.m, e.e, e.b, prec=ctx.p)
-        return MPMF._round_to_context(x, ctx)
+        d = Float._round_to_context(x, ctx)
+        return float(d)
+
+    def _visit_unknown(self, e: UnknownCall, ctx: EvalCtx):
+        raise NotImplementedError('unknown call', e)
 
     def _apply_method(self, e: NaryExpr, ctx: EvalCtx):
         fn = _method_table[e.name]
-        args: list[Digital] = []
+        args: list[float] = []
         for arg in e.children:
             val = self._visit_expr(arg, ctx)
-            if not isinstance(val, Digital):
+            if not isinstance(val, float):
                 raise TypeError(f'expected a real number argument for {e.name}, got {val}')
             args.append(val)
-
-        # compute the result
-        print(list(map(float, args)))
-        ctx = self._eval_ctx(ctx)
         try:
-            result = fn(*args, ctx=ctx)
-            print(result)
-        except gmpmath.SignedOverflow as e:
-            # we overflowed beyond MPFR's limits, generate a large value and round it
-            exp = ctx.emax + 1
-            x = Digital(negative=e.sign, c=1, exp=exp)
-            result = MPMF._round_to_context(x, ctx=ctx)
-        except gmpmath.SignedUnderflow as e:
-            # we underflowed beyond MPFR's limits, generate a small value and round it
-            ctx = self._eval_ctx(ctx)
-            exp = ctx.emin - ctx.p - 1
-            x = Digital(negative=e.sign, c=1, exp=exp)
-            result = MPMF._round_to_context(x, ctx=ctx)
+            result = fn(*args)
+        except OverflowError:
+            # We could return an infinity, but we don't know which one
+            result = math.nan
+        except ValueError:
+            # domain error means NaN
+            result = math.nan
 
         return result
 
     def _apply_cast(self, e: Cast, ctx: EvalCtx):
         x = self._visit_expr(e.children[0], ctx)
-        if not isinstance(x, Digital):
-            raise TypeError(f'expected a real number argument, got {x}')
-        ctx = self._eval_ctx(ctx)
-        return MPMF._round_to_context(x, ctx)
+        if not isinstance(x, float):
+            raise TypeError(f'expected a float, got {x}')
+        return x
 
     def _apply_not(self, e: Not, ctx: EvalCtx):
         arg = self._visit_expr(e.children[0], ctx)
@@ -283,15 +255,17 @@ class _Interpreter(ReduceVisitor):
 
     def _apply_range(self, e: Range, ctx: EvalCtx):
         stop = self._visit_expr(e.children[0], ctx)
-        if not isinstance(stop, Digital):
+        if not isinstance(stop, float):
             raise TypeError(f'expected a real number argument, got {stop}')
         if not stop.is_integer():
             raise TypeError(f'expected an integer argument, got {stop}')
-        return NDArray([MPMF(i, ctx) for i in range(int(stop))])
+        return tuple([float(i) for i in range(int(stop))])
 
     def _visit_nary_expr(self, e: NaryExpr, ctx: EvalCtx):
         if e.name in _method_table:
             return self._apply_method(e, ctx)
+        elif e.name == 'fma':
+            raise NotImplementedError('fma not supported in Python 3.11')
         elif isinstance(e, Cast):
             return self._apply_cast(e, ctx)
         elif isinstance(e, Not):
@@ -304,19 +278,6 @@ class _Interpreter(ReduceVisitor):
             return self._apply_range(e, ctx)
         else:
             raise NotImplementedError('unknown n-ary expression', e)
-
-    def _visit_unknown(self, e: UnknownCall, ctx: EvalCtx):
-        args = [self._visit_expr(arg, ctx) for arg in e.children]
-        fn = self.foreign[e.name]
-        if isinstance(fn, Function):
-            # calling FPy function
-            rt = _Interpreter(fn.env, override_ctx=self.override_ctx)
-            return rt.eval(fn.ir, args, ctx)
-        elif callable(fn):
-            # calling foreign function
-            return fn(*args)
-        else:
-            raise RuntimeError(f'not a function {fn}')
 
     def _apply_cmp2(self, op: CompareOp, lhs, rhs):
         match op:
@@ -345,42 +306,26 @@ class _Interpreter(ReduceVisitor):
         return True
 
     def _visit_tuple_expr(self, e: TupleExpr, ctx: EvalCtx):
-        return NDArray([self._visit_expr(x, ctx) for x in e.children])
+        return tuple([self._visit_expr(x, ctx) for x in e.children])
 
     def _visit_tuple_ref(self, e: TupleRef, ctx: EvalCtx):
         value = self._visit_expr(e.value, ctx)
-        if not isinstance(value, NDArray):
+        if not isinstance(value, tuple):
             raise TypeError(f'expected a tensor, got {value}')
 
-        slices: list[int] = []
+        elt = value
         for s in e.slices:
             val = self._visit_expr(s, ctx)
             if not isinstance(val, Digital):
                 raise TypeError(f'expected a real number slice, got {val}')
             if not val.is_integer():
                 raise TypeError(f'expected an integer slice, got {val}')
-            slices.append(int(val))
+            elt = elt[int(val)]
 
-        return value[slices]
+        return elt
 
     def _visit_tuple_set(self, e: TupleSet, ctx: EvalCtx):
-        value = self._visit_expr(e.array, ctx)
-        if not isinstance(value, NDArray):
-            raise TypeError(f'expected a tensor, got {value}')
-        value = NDArray(value) # make a copy
-
-        slices: list[int] = []
-        for s in e.slices:
-            val = self._visit_expr(s, ctx)
-            if not isinstance(val, Digital):
-                raise TypeError(f'expected a real number slice, got {val}')
-            if not val.is_integer():
-                raise TypeError(f'expected an integer slice, got {val}')
-            slices.append(int(val))
-
-        val = self._visit_expr(e.value, ctx)
-        value[slices] = val
-        return value
+        raise NotImplementedError
 
     def _apply_comp(
         self,
@@ -394,7 +339,7 @@ class _Interpreter(ReduceVisitor):
         else:
             var, iterable = bindings[0]
             array = self._visit_expr(iterable, ctx)
-            if not isinstance(array, NDArray):
+            if not isinstance(array, tuple):
                 raise TypeError(f'expected a tensor, got {array}')
             for val in array:
                 if isinstance(var, NamedId):
@@ -411,8 +356,9 @@ class _Interpreter(ReduceVisitor):
         for var in e.vars:
             if isinstance(var, NamedId):
                 del self.env[var]
-
-        return NDArray(elts)
+ 
+        # the result
+        return tuple(elts)
 
     def _visit_if_expr(self, e: IfExpr, ctx: EvalCtx):
         cond = self._visit_expr(e.cond, ctx)
@@ -420,12 +366,8 @@ class _Interpreter(ReduceVisitor):
             raise TypeError(f'expected a boolean, got {cond}')
         return self._visit_expr(e.ift if cond else e.iff, ctx)
 
-    def _visit_var_assign(self, stmt: VarAssign, ctx: EvalCtx) -> None:
+    def _visit_var_assign(self, stmt: VarAssign, ctx: EvalCtx):
         val = self._visit_expr(stmt.expr, ctx)
-        if self.enable_trace:
-            entry = ExprTraceEntry(stmt.expr, val, dict(self.env), ctx)
-            self.trace.append(entry)
-
         match stmt.var:
             case NamedId():
                 self.env[stmt.var] = val
@@ -434,7 +376,7 @@ class _Interpreter(ReduceVisitor):
             case _:
                 raise NotImplementedError('unknown variable', stmt.var)
 
-    def _unpack_tuple(self, binding: TupleBinding, val: NDArray, ctx: EvalCtx) -> None:
+    def _unpack_tuple(self, binding: TupleBinding, val: tuple, ctx: EvalCtx) -> None:
         if len(binding.elts) != len(val):
             raise NotImplementedError(f'unpacking {len(val)} values into {len(binding.elts)}')
         for elt, v in zip(binding.elts, val):
@@ -448,34 +390,14 @@ class _Interpreter(ReduceVisitor):
                 case _:
                     raise NotImplementedError('unknown tuple element', elt)
 
-    def _visit_tuple_assign(self, stmt: TupleAssign, ctx: EvalCtx) -> None:
+    def _visit_tuple_assign(self, stmt: TupleAssign, ctx: EvalCtx):
         val = self._visit_expr(stmt.expr, ctx)
-        if not isinstance(val, NDArray):
+        if not isinstance(val, tuple):
             raise TypeError(f'expected a tuple, got {val}')
-
-        if self.enable_trace:
-            entry = ExprTraceEntry(stmt.expr, val, dict(self.env), ctx)
-            self.trace.append(entry)
-
         self._unpack_tuple(stmt.binding, val, ctx)
 
-    def _visit_ref_assign(self, stmt: RefAssign, ctx: EvalCtx) -> None:
-        # lookup array
-        array = self._lookup(stmt.var)
-
-        # evaluate indices
-        slices: list[int] = []
-        for s in stmt.slices:
-            val = self._visit_expr(s, ctx)
-            if not isinstance(val, Digital):
-                raise TypeError(f'expected a real number slice, got {val}')
-            if not val.is_integer():
-                raise TypeError(f'expected an integer slice, got {val}')
-            slices.append(int(val))
-
-        # evaluate and update array
-        val = self._visit_expr(stmt.expr, ctx)
-        array[slices] = val
+    def _visit_ref_assign(self, stmt: RefAssign, ctx: EvalCtx):
+        raise NotImplementedError
 
     def _visit_if1_stmt(self, stmt: If1Stmt, ctx: EvalCtx):
         cond = self._visit_expr(stmt.cond, ctx)
@@ -485,29 +407,28 @@ class _Interpreter(ReduceVisitor):
             self._visit_block(stmt.body, ctx)
             for phi in stmt.phis:
                 self.env[phi.name] = self.env[phi.rhs]
+                del self.env[phi.rhs]
         else:
             for phi in stmt.phis:
                 self.env[phi.name] = self.env[phi.lhs]
+                del self.env[phi.lhs]
 
-    def _visit_if_stmt(self, stmt: IfStmt, ctx: EvalCtx) -> None:
+    def _visit_if_stmt(self, stmt: IfStmt, ctx: EvalCtx):
         cond = self._visit_expr(stmt.cond, ctx)
         if not isinstance(cond, bool):
             raise TypeError(f'expected a boolean, got {cond}')
-
-        if self.enable_trace:
-            entry = ExprTraceEntry(stmt.cond, cond, dict(self.env), ctx)
-            self.trace.append(entry)
-
-        if cond:
+        elif cond:
             self._visit_block(stmt.ift, ctx)
             for phi in stmt.phis:
                 self.env[phi.name] = self.env[phi.lhs]
+                del self.env[phi.lhs]
         else:
             self._visit_block(stmt.iff, ctx)
             for phi in stmt.phis:
                 self.env[phi.name] = self.env[phi.rhs]
+                del self.env[phi.rhs]
 
-    def _visit_while_stmt(self, stmt: WhileStmt, ctx: EvalCtx) -> None:
+    def _visit_while_stmt(self, stmt: WhileStmt, ctx: EvalCtx):
         for phi in stmt.phis:
             self.env[phi.name] = self.env[phi.lhs]
             del self.env[phi.lhs]
@@ -515,10 +436,6 @@ class _Interpreter(ReduceVisitor):
         cond = self._visit_expr(stmt.cond, ctx)
         if not isinstance(cond, bool):
             raise TypeError(f'expected a boolean, got {cond}')
-
-        if self.enable_trace:
-            entry = ExprTraceEntry(stmt.cond, cond, dict(self.env), ctx)
-            self.trace.append(entry)
 
         while cond:
             self._visit_block(stmt.body, ctx)
@@ -530,18 +447,13 @@ class _Interpreter(ReduceVisitor):
             if not isinstance(cond, bool):
                 raise TypeError(f'expected a boolean, got {cond}')
 
-            if self.enable_trace:
-                entry = ExprTraceEntry(stmt.cond, cond, dict(self.env), ctx)
-                self.trace.append(entry)
-
-
-    def _visit_for_stmt(self, stmt: ForStmt, ctx: EvalCtx) -> None:
+    def _visit_for_stmt(self, stmt: ForStmt, ctx: EvalCtx):
         for phi in stmt.phis:
             self.env[phi.name] = self.env[phi.lhs]
             del self.env[phi.lhs]
 
         iterable = self._visit_expr(stmt.iterable, ctx)
-        if not isinstance(iterable, NDArray):
+        if not isinstance(iterable, tuple):
             raise TypeError(f'expected a tensor, got {iterable}')
 
         for val in iterable:
@@ -559,8 +471,9 @@ class _Interpreter(ReduceVisitor):
                 props[k] = self._lookup(v)
             else:
                 props[k] = v
-
-        ctx = determine_ctx(ctx, props)
+        ctx = determine_ctx(ctx, stmt.props)
+        if not self._is_python_ctx(ctx):
+            raise NotImplementedError(f'unsupported context {ctx}')
         return self._visit_block(stmt.body, ctx)
 
     def _visit_assert(self, stmt: AssertStmt, ctx: EvalCtx):
@@ -571,17 +484,12 @@ class _Interpreter(ReduceVisitor):
             raise AssertionError(stmt.msg)
         return ctx
 
-    def _visit_effect(self, stmt: EffectStmt, ctx: EvalCtx):
+    def _visit_effect(self, stmt, ctx):
         self._visit_expr(stmt.expr, ctx)
         return ctx
 
     def _visit_return(self, stmt: Return, ctx: EvalCtx):
-        val = self._visit_expr(stmt.expr, ctx)
-        if self.enable_trace:
-            entry = ExprTraceEntry(stmt.expr, val, dict(self.env), ctx)
-            self.trace.append(entry)
-
-        return val
+        return self._visit_expr(stmt.expr, ctx)
 
     def _visit_phis(self, phis, lctx, rctx):
         raise NotImplementedError('do not call directly')
@@ -599,27 +507,16 @@ class _Interpreter(ReduceVisitor):
     def _visit_function(self, func: FunctionDef, ctx: EvalCtx):
         raise NotImplementedError('do not call directly')
 
-    # override typing hint
-    def _visit_statement(self, stmt, ctx: EvalCtx) -> None:
-        return super()._visit_statement(stmt, ctx)
 
 
-class TitanicInterpreter(Interpreter):
+class PythonInterpreter(Interpreter):
     """
-    Standard interpreter for FPy programs.
+    Python-backed interpreter for FPy programs.
 
-    Programs are evaluated using the Titanic backend (`titanfp`).
-    Booleans are Python `bool` values, real numbers are Titanic `MPMF` values,
-    and tensors are Titanic `NDArray` values.
-
-    All operations are correctly-rounded.
+    Programs are evaluated using Python's `math` library.
+    Booleans are Python `bool` values, real numbers are `float` values,
+    and tensors are Python `tuple` values.
     """
-
-    ctx: Optional[EvalCtx] = None
-    """optionaly overriding context"""
-
-    def __init__(self, ctx: Optional[EvalCtx] = None):
-        self.ctx = ctx
 
     def eval(
         self,
@@ -629,14 +526,11 @@ class TitanicInterpreter(Interpreter):
     ):
         if not isinstance(func, Function):
             raise TypeError(f'Expected Function, got {func}')
-        rt = _Interpreter(func.env, override_ctx=self.ctx)
+        rt = _Interpreter(func.env)
         return rt.eval(func.ir, args, ctx)
 
     def eval_with_trace(self, func: Function, args: Sequence[Any], ctx = None):
-        rt = _Interpreter(func.env, override_ctx=self.ctx, enable_trace=True)
-        result = rt.eval(func.ir, args, ctx)
-        return result, rt.trace
+        raise NotImplementedError('not implemented')
 
     def eval_expr(self, expr: Expr, env: _Env, ctx: EvalCtx):
-        rt = _Interpreter(ForeignEnv.empty(), override_ctx=self.ctx, env=env)
-        return rt._visit_expr(expr, ctx)
+        raise NotImplementedError('not implemented')
