+++ fpy2/runtime/profile/expr_profiler.py
@@ -12,7 +12,7 @@ from titanfp.arithmetic.mpmf import MPMF
 from ..common import ExprTraceEntry
 from ..function import Function, Interpreter, get_default_interpreter
 from ..metric import ordinal_error
-from ..real.interpreter import RealInterpreter
+from ...interpret.real import RealInterpreter
 from ..real.rival_manager import PrecisionLimitExceeded
 
 from ...ir import Expr
