+++ fpy2/runtime/profile/function_profiler.py
@@ -11,7 +11,7 @@ from titanfp.arithmetic.mpmf import MPMF
 
 from ..function import Function, Interpreter, get_default_interpreter
 from ..metric import ordinal_error
-from ..real.interpreter import RealInterpreter
+from ...interpret.real import RealInterpreter
 from ..real.rival_manager import PrecisionLimitExceeded
 
 class FunctionProfileResult:
