+++ fpy2/runtime/real/__init__.py
@@ -1,2 +1,2 @@
-from .interpreter import RealInterpreter
+from ...interpret.real import RealInterpreter
 from .interval import Interval, BoolInterval, RealInterval
