+++ fpy2/runtime/titanic/__init__.py
@@ -1,1 +1,1 @@
-from .interpreter import TitanicInterpreter
+from ...interpret.titanic import TitanicInterpreter
